pyproject.toml
[project]
name = "eitpose"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12, <3.13"
dependencies = [
    "bottleneck>=1.5.0",
    "certifi>=2025.8.3",
    "cycler>=0.12.1",
    "fonttools>=4.59.1",
    "joblib>=1.5.1",
    "kiwisolver>=1.4.9",
    "matplotlib>=3.10.5",
    "mediapipe>=0.10.14",
    "mkl-service>=2.5.2",
    "munkres>=1.1.4",
    "numexpr>=2.11.0",
    "numpy>=2.3.2",
    "opencv-python>=4.11.0.86",
    "packaging>=25.0",
    "pandas>=2.3.1",
    "pillow>=11.3.0",
    "pip>=25.2",
    "plotly>=6.3.0",
    "ply>=3.11",
    "pyopengl>=3.1.9",
    "pyparsing>=3.2.3",
    "pyqt5>=5.15.11",
    "pyqtgraph>=0.13.7",
    "pyserial>=3.5",
    "pytz>=2025.2",
    "scikit-learn>=1.7.1",
    "scipy>=1.16.1",
    "setuptools>=80.9.0",
    "sip>=6.12.0",
    "six>=1.17.0",
    "tomli>=2.2.1",
    "tornado>=6.5.2",
    "tzdata>=2025.2",
    "wheel>=0.45.1",
]

[dependency-groups]
dev = []
